CAT-2150 Firebase Exception Reporting

+++ catroid/src/main/java/org/catrobat/catroid/io/StorageHandler.java
@@ -31,6 +31,7 @@ import android.util.Log;
 
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
+import com.google.firebase.crash.FirebaseCrash;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonIOException;
@@ -277,6 +278,7 @@ public final class StorageHandler {
 		try {
 			INSTANCE = new StorageHandler();
 		} catch (IOException ioException) {
+			FirebaseCrash.report(ioException);
 			throw new RuntimeException("Initialize StorageHandler failed");
 		}
 	}
@@ -343,11 +345,13 @@ public final class StorageHandler {
 			}
 			outputStream.flush();
 		} catch (IOException ioException) {
+			FirebaseCrash.report(ioException);
 			Log.e(TAG, Log.getStackTraceString(ioException));
 		} finally {
 			try {
 				outputStream.close();
 			} catch (IOException e) {
+				FirebaseCrash.report(e);
 				Log.e(TAG, "Could not close outputstream.", e);
 			}
 		}
@@ -542,6 +546,7 @@ public final class StorageHandler {
 		try {
 			createBackPackFileStructure();
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Creating backpack file structure failed");
 		}
 	}
@@ -558,6 +563,7 @@ public final class StorageHandler {
 				return null;
 			}
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Exception getFirstSceneName", e);
 			return null;
 		}
@@ -566,6 +572,7 @@ public final class StorageHandler {
 		try {
 			projectXml = Files.toString(projectXmlFile, Charsets.UTF_8);
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Exception getFirstSceneName", e);
 			return null;
 		}
@@ -587,6 +594,7 @@ public final class StorageHandler {
 				return loadSupportProject(projectName, context);
 			}
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Could not check Scene Tag!", e);
 			return null;
 		}
@@ -606,6 +614,7 @@ public final class StorageHandler {
 				project.getSceneByName(sceneName).getDataContainer().setProject(project);
 			}
 		} catch (Exception e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Could not get Project from xml and get Scene from order", e);
 			loadSaveLock.unlock();
 			return null;
@@ -681,6 +690,7 @@ public final class StorageHandler {
 			fixFolderStructureForSupportProject(projectName, project.getDefaultScene().getName());
 			return project;
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.d(TAG, "Could not load project!");
 			UtilFile.deleteDirectory(file);
 			Log.d(TAG, "loadProject: directory is deleted and "
@@ -692,6 +702,7 @@ public final class StorageHandler {
 				try {
 					fileInputStream.close();
 				} catch (IOException ioException) {
+					FirebaseCrash.report(ioException);
 					Log.e(TAG, "can't close fileStream.", ioException);
 				}
 			}
@@ -705,6 +716,7 @@ public final class StorageHandler {
 				fileInputStream.close();
 				return true;
 			} catch (IOException ioException) {
+				FirebaseCrash.report(ioException);
 				Log.e(TAG, "can't close fileStream.", ioException);
 			}
 		}
@@ -745,6 +757,7 @@ public final class StorageHandler {
 					Log.d(TAG, "Project version differ <" + oldProjectXml.length() + "> <"
 							+ projectXml.length() + ">. update " + currentCodeFile.getName());
 				} catch (Exception exception) {
+					FirebaseCrash.report(exception);
 					Log.e(TAG, "Opening old project " + currentCodeFile.getAbsolutePath() + " failed.", exception);
 					return false;
 				}
@@ -764,9 +777,9 @@ public final class StorageHandler {
 				writer.newLine();
 			}
 			writer.flush();
-
 			return true;
 		} catch (Exception exception) {
+			FirebaseCrash.report(exception);
 			Log.e(TAG, "Saving project " + project.getName() + " failed.", exception);
 			return false;
 		} finally {
@@ -782,6 +795,7 @@ public final class StorageHandler {
 						Log.e(TAG, "Could not rename " + currentCodeFile.getName());
 					}
 				} catch (IOException ioException) {
+					FirebaseCrash.report(ioException);
 					Log.e(TAG, "Failed closing the buffered writer", ioException);
 				}
 			}
@@ -823,6 +837,7 @@ public final class StorageHandler {
 			ProjectManager.getInstance().setCurrentScene(currentProject.getDefaultScene());
 			return project.getDefaultScene();
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Error while creating default Scene!", e);
 			return null;
 		}
@@ -843,6 +858,7 @@ public final class StorageHandler {
 			writer.write(json);
 			return true;
 		} catch (IOException e) {
+			FirebaseCrash.report(e);
 			Log.e(TAG, "Could not write backpack file", e);
 			return false;
 		} finally {
@@ -850,6 +866,7 @@ public final class StorageHandler {
 				try {
 					writer.close();
 				} catch (IOException ioException) {
+					FirebaseCrash.report(ioException);
 					Log.e(TAG, "Failed closing the buffered writer", ioException);
 				}
 			}
@@ -868,9 +885,11 @@ public final class StorageHandler {
 			BufferedReader bufferedBackpackReader = new BufferedReader(new FileReader(backpackFile));
 			return backpackGson.fromJson(bufferedBackpackReader, Backpack.class);
 		} catch (FileNotFoundException e) {
+			FirebaseCrash.report(e);
 			Log.d(TAG, "Could not find backpack file!");
 			return new Backpack();
 		} catch (JsonSyntaxException | JsonIOException jsonException) {
+			FirebaseCrash.report(jsonException);
 			Log.d(TAG, "Could not load backpack file! File will be deleted!", jsonException);
 			deleteBackpackFile();
 			return new Backpack();
@@ -916,6 +935,7 @@ public final class StorageHandler {
 				}
 			}
 		} catch (Exception exception) {
+			FirebaseCrash.report(exception);
 			Log.e(TAG, "Exception " + exception);
 		} finally {
 			loadSaveLock.unlock();
@@ -1242,6 +1262,7 @@ public final class StorageHandler {
 				toDelete.delete();
 			}
 		} catch (FileNotFoundException fileNotFoundException) {
+			FirebaseCrash.report(fileNotFoundException);
 			Log.e(TAG, Log.getStackTraceString(fileNotFoundException));
 		}
 	}
@@ -1351,18 +1372,25 @@ public final class StorageHandler {
 		}
 
 		File sourceDirectory = new File(buildPath(buildScenePath(sourceProject, sourceScene), type));
-		File targetDirectory = new File(buildPath(buildScenePath(targetProject, targetScene), type));
-		targetDirectory.mkdirs();
 
-		for (File sourceFile : sourceDirectory.listFiles()) {
-			File targetFile = new File(targetDirectory.getAbsolutePath(), sourceFile.getName());
-			targetFile.createNewFile();
+		if (!targetDirectory.exists() || !sourceDirectory.exists()) {
+			return false;
+		}
+		try {
+			for (File sourceFile : sourceDirectory.listFiles()) {
+			  File targetFile = new File(targetDirectory.getAbsolutePath(), sourceFile.getName());
+			  targetFile.createNewFile();
 
-			FileChannel source = new FileInputStream(sourceFile).getChannel();
-			FileChannel target = new FileOutputStream(targetFile).getChannel();
-			target.transferFrom(source, 0, source.size());
-			source.close();
-			target.close();
+			  FileChannel source = new FileInputStream(sourceFile).getChannel();
+			  FileChannel target = new FileOutputStream(targetFile).getChannel();
+			  target.transferFrom(source, 0, source.size());
+			  source.close();
+			  target.close();
+			}
+		} catch (IOException e) {
+			FirebaseCrash.report(e);
+			Log.e(TAG, e.getMessage());
+			return false;
 		}
 	}
